requirements.txt
# Core
numpy
gymnasium==0.26.2
mujoco==3.1.2
myosuite==2.8.4

# Docs / parsing / configs
pyyaml
jsonschema

# Rendering / images
imageio

# Dev / testing
pytest